Update PetOwnerController, PetOwnerService

+++ src/main/java/pro/sky/animal_shelter_telegram_bot/service/impl/PetOwnerServiceImpl.java
@@ -37,23 +37,35 @@ public class PetOwnerServiceImpl implements PetOwnerService {
     }
 
     @Override
-    public void deletePetOwner(Long id) {
+    public boolean deletePetOwner(Long id) {
+        if (petOwnerRepository.findById(id).isEmpty()){
+            logger.info("Pet with id {} is not found", id);
+            return false;
+        }
         petOwnerRepository.deleteById(id);
-        logger.info("Pet owner with id {} is deleted", id);
-
+        logger.info("Pet with id {} is deleted", id);
+        return true;
     }
 
     @Override
     public PetOwner findPetOwner(Long id) {
-        PetOwner findingPetOwner = petOwnerRepository.findById(id).get();
-        logger.info("Pet owner with id {} is found", id);
-        return findingPetOwner;
+        if (petOwnerRepository.findById(id).isEmpty()){
+            logger.info("Pet with id {} is not found", id);
+            return null;
+        }
+        PetOwner petOwner = petOwnerRepository.findById(id).get();
+        logger.info("Pet with id {} is found", id);
+        return petOwner;
     }
 
     @Override
     public PetOwner changePetOwner(PetOwner petOwner) {
+        if (petOwnerRepository.findById(petOwner.getId()).isEmpty()){
+            logger.info("Pet with id {} is not found", petOwner.getId());
+            return null;
+        }
         PetOwner changingPetOwner = petOwnerRepository.save(petOwner);
-        logger.info("Pet owner {} is saved", petOwner);
+        logger.info("Pet {} is saved", petOwner);
         return changingPetOwner;
     }
